requirements.txt
autopep8==1.6.0
Backtesting==0.3.3
bokeh==2.4.3
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.33.3
idna==3.3
Jinja2==3.1.2
kiwisolver==1.4.3
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib==3.5.2
multitasking==0.0.10
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pycodestyle==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
six==1.16.0
TA-Lib==0.4.24
toml==0.10.2
tornado==6.1
typing_extensions==4.2.0
urllib3==1.26.9
yfinance==0.1.70
